[init] Fix a compile error with Config::get<int>; use uint32_t instead.

+++ src/init.cpp
@@ -872,7 +872,7 @@ void SimInit(const char* configFile, const char* outputDir, uint32_t shmid) {
     //NOTE: This should be as early as possible, so that we can attach to the debugger before initialization.
     zinfo->attachDebugger = config.get<bool>("sim.attachDebugger", false);
     zinfo->harnessPid = getppid();
-    zinfo->debugPortId = config.get<int>("sim.debugPortId", 0);
+    zinfo->debugPortId = static_cast<int>(config.get<uint32_t>("sim.debugPortId", 0));
     getLibzsimAddrs(&zinfo->libzsimAddrs);
 
     if (zinfo->attachDebugger) {
